Finalisation du système d'authentification et changement du style

+++ app.py
@@ -1,7 +1,12 @@
 import sqlite3
-from flask import Flask, render_template, request, url_for, flash, redirect
+from flask import *
 from werkzeug.exceptions import abort
 from hashlib import sha256
+from http import cookies
+import secrets
+
+app = Flask(__name__)
+app.config['SECRET_KEY'] = 'your secret key'
 
 def get_db_connection():
     conn = sqlite3.connect('database.db')
@@ -17,12 +22,21 @@ def get_post(post_id):
     if post is None:
         abort(404)
     return post
-app = Flask(__name__)
-app.config['SECRET_KEY'] = 'your secret key'
 
+def is_authenticated():
+    token = request.cookies.get('token')
+    if token:
+        conn = get_db_connection()
+        user = conn.execute('SELECT * FROM user WHERE token = ?', (token,)).fetchone()
+        conn.close()
+        if user :
+            return True
+    return False
 
 @app.route('/')
 def index():
+    if not is_authenticated():
+        return redirect(url_for('user_login'))
     conn = get_db_connection()
     posts = conn.execute('SELECT * FROM posts').fetchall()
     conn.close()
@@ -34,12 +48,17 @@ def index():
 
 @app.route('/post/<int:post_id>')
 def post_get(post_id):
+    if not is_authenticated():
+        return redirect(url_for('user_login'))
     post = get_post(post_id)
     return render_template('post/get.html', post=post)
 
 
 @app.route('/post/create', methods=('GET', 'POST'))
 def post_create():
+    if not is_authenticated():
+        return redirect(url_for('user_login'))
+
     if request.method == 'POST':
         title = request.form['title']
         content = request.form['content']
@@ -59,8 +78,9 @@ def post_create():
 
 @app.route('/post/<int:id>/edit', methods=('GET', 'POST'))
 def post_edit(id):
+    if not is_authenticated():
+        return redirect(url_for('user_login'))
     post = get_post(id)
-
     if request.method == 'POST':
         title = request.form['title']
         content = request.form['content']
@@ -80,6 +100,8 @@ def post_edit(id):
 
 @app.route('/post/<int:id>/delete', methods=('POST',))
 def post_delete(id):
+    if not is_authenticated():
+        return redirect(url_for('user_login'))
     post = get_post(id)
     conn = get_db_connection()
     conn.execute('DELETE FROM posts WHERE id = ?', (id,))
@@ -88,8 +110,15 @@ def post_delete(id):
     flash('"{}" was successfully deleted!'.format(post['title']))
     return redirect(url_for('index'))
 
+##
+## USER
+##
+
 @app.route('/user/login', methods=('GET', 'POST'))
 def user_login():
+    if is_authenticated():
+        return redirect(url_for('index'))
+
     if request.method == 'POST':
         ## information tirer du html de ce que l'utilisateur a ecris
         username = request.form['username']
@@ -106,7 +135,14 @@ def user_login():
             conn.close()
             ## verifation de user si il est vide alors erreur si il est rempli alor ca marche
             if user:
-                return redirect(url_for('index'))
+                token = secrets.token_urlsafe(64)
+                conn = get_db_connection()
+                conn.execute('UPDATE user SET token = ? WHERE id = ?',(token, user["id"]))
+                conn.commit()
+                conn.close()
+                response = make_response(redirect(url_for('index')))
+                response.set_cookie( "token", token )
+                return response
             else:
                 flash('Username ou mot de passe incorrect')
